ImplementsInterfaces to Generator added
git-svn-id: svn+ssh://dasz.at/srv/dasz/svn/Kistl@924 76cbe2a5-d633-47ad-8992-a0d0e66ffbf2

+++ Kistl.Server/Generators/FrozenObjects/Implementation/ObjectClasses/DataStore.cs
@@ -53,7 +53,7 @@ namespace Kistl.Server.Generators.FrozenObjects.Implementation.ObjectClasses
                 if (prop is EnumerationProperty)
                 {
                     var enumProp = (EnumerationProperty)prop;
-                    return string.Format("({0}){1}", enumProp.Enumeration.ClassName, value);
+                    return string.Format("({0}){1}", enumProp.Enumeration.ClassName, (int)value);
                 }
                 else if (prop is ValueTypeProperty)
                 {
